move expand_tilde back to shell, but make it accessible to sl_compiler
