+++ shell/Cargo.toml
@@ -10,7 +10,6 @@ sl-liner = { git = "https://github.com/sl-sh-dev/sl-liner.git" }
 nix = { version = "0.27.1", features = ["signal", "process", "term", "user", "fs", "hostname"] }
 glob = "0.3"
 cfg-if = "1.0"
-builtins = { path = "../builtins" }
 
 [build-dependencies]
 chrono = "0.4.7"
